Remove time-based mode
If using sleep in the master's node (panel id 0), there's no need for a 
time-based mode.

Things will get a bit easier.

If a command is sent only to node1, node2, ... but not the master, the 
script will get stuck. To prevent this, a "rc" call is sent to the 
master (panel id 0).